CCTools: Add Deterministic Bucketing Algorithms (det-greedy, det-exhaust) (#4302)
* add devs for det greedy bucketing

* done adding algorithms

* add tests for bucketing

* add tests for the bucketing manager

* refactor exhaust bucketing algos

* refactor greedy algos

* implement det-greedy and det-exhaust

* format for lint

* fix test

* fix test

* lint

* small fixes

* trigger rerun github actions

* add Copilot fixes

* lint

* lint

* try 1s monitor interval

* try factory period of 10s

* try 15s monitor interval

* try 60s monitor interval

* try 300s interval

* add envlist support to wq factory

* revert monitor interval change

* merge resource report instead of replace

* replace instead of merge when success

* add simulate, result not good

* add simulation code

* fix bucketing algos bug

* fix bug where resources are not predicted correctly

* fix resource counting in taskvine

* fix bucketing manager how to predict and add res report

* save fix resource int to double

* modify logic in task exceed limits but not this resource to take into account whether it is in the sampling phase

* move max function to top of bucketing manager

* rm tmp.log

* disable proportional resources in source code

* improve sim topeft code

* add quantized bucketing

* change num buckets of quantized from 2 to 1 for blind info

* add quantized to topeft simulation

* explicitly disable proportional resources in src code

* fix warning compile

* fix wq_factory

* fix bucket manager

* remove run_sim

* remove run sim from gitignore

* remove run_sim

* remove custom wq fix

* add proportional whole tasks back

* lint

---------

+++ taskvine/src/worker/vine_worker.c
@@ -1513,13 +1513,13 @@ static int enforce_worker_limits(struct link *manager)
 {
 	if (options->disk_total > 0 && total_resources->disk.inuse > options->disk_total) {
 		fprintf(stderr,
-				"vine_worker: %s used more than declared disk space (--disk - < disk used) %" PRIu64 " < %" PRIu64 " MB\n",
+				"vine_worker: %s used more than declared disk space (--disk - < disk used) %" PRIu64 " < %lf MB\n",
 				workspace->workspace_dir,
 				options->disk_total,
 				total_resources->disk.inuse);
 
 		if (manager) {
-			send_message(manager, "info disk_exhausted %lld\n", (long long)total_resources->disk.inuse);
+			send_message(manager, "info disk_exhausted %lf\n", total_resources->disk.inuse);
 		}
 
 		return 0;
@@ -1527,12 +1527,12 @@ static int enforce_worker_limits(struct link *manager)
 
 	if (options->memory_total > 0 && total_resources->memory.inuse > options->memory_total) {
 		fprintf(stderr,
-				"vine_worker: used more than declared memory (--memory < memory used) %" PRIu64 " < %" PRIu64 " MB\n",
+				"vine_worker: used more than declared memory (--memory < memory used) %" PRIu64 " < %lf MB\n",
 				options->memory_total,
 				total_resources->memory.inuse);
 
 		if (manager) {
-			send_message(manager, "info memory_exhausted %lld\n", (long long)total_resources->memory.inuse);
+			send_message(manager, "info memory_exhausted %lf\n", total_resources->memory.inuse);
 		}
 
 		return 0;
@@ -1557,12 +1557,12 @@ static int enforce_worker_promises(struct link *manager)
 
 	if (options->disk_total > 0 && total_resources->disk.total < options->disk_total) {
 		fprintf(stderr,
-				"vine_worker: has less than the promised disk space (--disk > disk total) %" PRIu64 " < %" PRIu64 " MB\n",
+				"vine_worker: has less than the promised disk space (--disk > disk total) %" PRIu64 " < %lf MB\n",
 				options->disk_total,
 				total_resources->disk.total);
 
 		if (manager) {
-			send_message(manager, "info disk_error %lld\n", (long long)total_resources->disk.total);
+			send_message(manager, "info disk_error %lf\n", total_resources->disk.total);
 		}
 
 		return 0;
@@ -2339,7 +2339,7 @@ int main(int argc, char *argv[])
 
 	/* Display the available resources once at startup. */
 	measure_worker_resources();
-	printf("vine_worker: using %" PRId64 " cores, %" PRId64 " MB memory, %" PRId64 " MB disk, %" PRId64 " gpus\n",
+	printf("vine_worker: using %lf cores, %lf MB memory, %lf MB disk, %lf gpus\n",
 			total_resources->cores.total,
 			total_resources->memory.total,
 			total_resources->disk.total,
